+++ CHANGELOG.md
@@ -78,6 +78,7 @@ Version 0.7.0-SNAPSHOT - Build 20160119
 - Changed the names of setters/getters of Wrapper classes.
 - Increased the test-coverage.
 - Performed profiling and changed thread logic where necessary to improve speed.
+- The TextClassifierTest no longer passes a prefix in each test.
 
 Version 0.6.1 - Build 20160102
 ------------------------------